Delete function in admin page for sme and cleantech users and alert message

+++ goeko/apps/goeko/src/app/admin/data-admin/admin-users.component.ts
@@ -1,4 +1,4 @@
-import { setRemoteDefinitions } from '@nx/angular/mf';
+
 import { CommonModule } from '@angular/common'
 import { Component } from '@angular/core'
 import { CleanTechService, SmeService, USER_TYPE, UserType } from '@goeko/store'
@@ -6,6 +6,7 @@ import { Observable } from 'rxjs'
 import { DATA_ACTOR_SWITCH } from '../data-actors-switch.constants'
 import { MessageService } from '@goeko/business-ui'
 import { MESSAGE_TYPE } from '@goeko/ui';
+
 
 interface User {
   id: number
@@ -21,7 +22,7 @@ type DataSourcesByUserType = {
   selector: 'goeko-data-admin',
   standalone: true,
   imports: [CommonModule],
-  providers: [SmeService, CleanTechService],
+  providers: [SmeService, CleanTechService, MessageService],
   templateUrl: './admin-users.component.html',
   styleUrl: './admin-users.component.scss',
 })
@@ -34,6 +35,7 @@ export class AdminUserComponent {
   public get dataSources() {
     return this._dataSourcesByUserType[this.selectedUserType as keyof DataSourcesByUserType] as Observable<User[]>
   }
+
   public headers: { title: string; key: keyof User }[] = [
     {
       title: 'NAME',
@@ -72,36 +74,25 @@ export class AdminUserComponent {
     this.selectedUserType = type
   }
   deleteUser(id: number): void {
-    console.log('delete user', id)
-    this.MessageService.deleteMessage(MESSAGE_TYPE.WARNING, `Usuario ${id}`).subscribe();
-      switch(this.selectedUserType) {
-        case USER_TYPE.SME:
-          this._smeServices.deleteSmeUser(id).subscribe(() => {
-            console.log(`Deleted SME user with ID ${id}`);
-              // Aquí puedes agregar cualquier lógica adicional después de la eliminación, como actualizar la lista de usuarios
-          }, error => {
-            console.error(`Error deleting SME user with ID ${id}: `, error);
-          });
-          break;
-        case USER_TYPE.CLEANTECH:
-          this._cleantechServices.deleteCleantechUser(id).subscribe(() => {
-            console.log(`Deleted Cleantech user with ID ${id}`);
-              // Aquí puedes agregar cualquier lógica adicional después de la eliminación, como actualizar la lista de usuarios
-          }, error => {
-            console.error(`Error deleting Cleantech user with ID ${id}: `, error);
-          });
-          break;
-        default:
-          console.error(`Unsupported user type: ${this.selectedUserType}`);
+    this.MessageService.deleteMessage(MESSAGE_TYPE.WARNING, `${id}`).afterClosed().subscribe((isConfirmed) => {
+      if (isConfirmed) {
+        switch(this.selectedUserType) {
+          case USER_TYPE.SME:
+            this._smeServices.deleteSmeUser(id).subscribe(() => {
+              this._smeServices.getAllSmesData().subscribe(smesData => {
+              });
+            });
+            break;
+          case USER_TYPE.CLEANTECH:
+            this._cleantechServices.deleteCleantechUser(id).subscribe(() => {
+              this._cleantechServices.getAllCleantechData().subscribe(cleantechData => {
+              });
+            });
+            break;
+          default:
+            console.error(`Unsupported user type: ${this.selectedUserType}`);
+        }
       }
-    }
-
-
-
-
-
-    ///this.adminService.deleteUser(id, selectedUserType)
-    // deleteUser(id, selectedUserType) {
-    //switch(selectedUserType) { case return this._smeServices.deleteSme(id); case return this._cleantechServices.deleteCleantech(id); }
-
+    });
+  }
 }
